move JobOrderData to bespin v2 joborder.py
